refactor: split star disk and ship lod modules
