+++ src/types/ai/state.ts
@@ -1,0 +1,174 @@
+import type { Vector3 } from 'three';
+import type { EntityId } from '../core.js';
+import type { Team, ShipHull } from '../gameplay.js';
+import type { DoctrineState } from './doctrine.js';
+import type { AIMetrics } from './metrics.js';
+
+export type AIIntent =
+  | 'Attack'
+  | 'Kite'
+  | 'Escort'
+  | 'Intercept'
+  | 'Flee'
+  | 'Regroup'
+  | 'Reposition';
+
+export interface AICommand {
+  heading: Vector3;
+  thrust: number;
+  /** Optional lateral strafe input in range [-1, 1]. */
+  strafe?: number;
+  firePrimary: boolean;
+  orbit?: number;
+  targetId?: EntityId;
+  ttl: number;
+}
+
+export interface AITraits {
+  aggression: number;
+  patience: number;
+  dodge: number;
+}
+
+export interface AIState {
+  profileId: string;
+  intent: AIIntent;
+  nextThinkAt: number;
+  cooldowns: {
+    dodgeAt: number;
+    burstAt: number;
+  };
+  lod: 0 | 1 | 2;
+  traitSeed: number;
+  traits: AITraits;
+  targetId?: EntityId;
+  lastScore?: number;
+  command: AICommand;
+  stickinessUntil: number;
+  stickinessHeading: Vector3;
+  stickinessTargetId?: EntityId;
+  desiredRange?: readonly [number, number];
+}
+
+export interface BehaviorProfile {
+  desiredRange: readonly [number, number];
+  orbit: number;
+  aggression: number;
+  patience: number;
+  dodgeFreq: number;
+  classBias: Partial<Record<ShipHull, number>>;
+  style: 'brawler' | 'kiter' | 'artillery' | 'escort';
+  gates?: {
+    ammoMin?: number;
+    hpRetreatPct?: number;
+  };
+  verticalManeuver: number;
+  elevationPreference?: 'above' | 'below' | 'follow';
+  bandPreference?: 'outer' | 'mid' | 'inner';
+  engagementBias?: number;
+}
+
+export type TeamPosture = 'aggressive' | 'hold' | 'retreat';
+
+export interface AIBlackboard {
+  tickIndex: number;
+  teamPosture: Record<Team, TeamPosture>;
+  allyCentroid: Record<Team, Vector3>;
+  nearestEnemy: Map<EntityId, EntityId>;
+  threatToVip: Map<EntityId, EntityId>;
+  tmpVectors: Vector3[];
+  strengthRatio: Record<Team, number>;
+  teamPriority: Record<Team, PrioritisedTarget[]>;
+  priorityIndex: Record<Team, Map<EntityId, number>>;
+  focusFire: Record<Team, Map<EntityId, number>>;
+  visibleEnemies?: Record<Team, Map<EntityId, SensorVisibility>>;
+  teamCounts?: Record<Team, number>;
+  // Vertical dispersion tracking for validation (optional for backward compatibility)
+  verticalDispersion?: {
+    headingYSamples: number[];
+    positionYSamples: number[];
+    lastUpdateTick: number;
+  };
+}
+
+export interface AITeamAssignments {
+  escorts: Map<EntityId, EscortAssignment>;
+}
+
+export interface EscortAssignment {
+  vipId: EntityId;
+  offset: Vector3;
+  threatId?: EntityId;
+}
+
+export interface AIIntentSnapshot {
+  tick: number;
+  time: number;
+  counts: Partial<Record<AIIntent, number>>;
+  total: number;
+}
+
+export interface PrioritisedTarget {
+  id: EntityId;
+  threat: number;
+  distanceSq: number;
+  focusLoad: number;
+}
+
+export type AIInterruptReason = 'hp-drop' | 'target-lost' | 'vip-threat' | 'manual';
+
+export interface IntentInterruptEvent {
+  shipId: EntityId;
+  reason: AIInterruptReason;
+  tick: number;
+  sourceId?: EntityId;
+}
+
+export interface AIInterruptState {
+  cooldownTick: Map<string, number>;
+  damageThisTick: Map<EntityId, number>;
+  lastDamageTick: number;
+  vipThreatAssignments: Map<EntityId, EntityId>;
+}
+
+export interface AIShotHistogram {
+  buckets: readonly number[];
+  counts: number[];
+  total: number;
+}
+
+export interface AIInBandStats {
+  samples: number;
+  satisfied: number;
+}
+
+export interface SensorVisibility {
+  strength: number;
+  lastSeenTick: number;
+  sourceId: EntityId;
+  occluded: boolean;
+  distance: number;
+}
+
+export interface SensorState {
+  lastUpdateTick: number;
+  visibilityByTeam: Record<Team, Map<EntityId, SensorVisibility>>;
+  decayRate: number;
+  threshold: number;
+  staleDecay: number;
+}
+
+export interface AIManagerState {
+  enabled: boolean;
+  tickInterval: number;
+  maxPerTick: number;
+  accumulator: number;
+  tickIndex: number;
+  cursor: number;
+  slices: number;
+  assignments: AITeamAssignments;
+  metrics: AIMetrics;
+  doctrine?: DoctrineState;
+  interrupts?: IntentInterruptEvent[];
+  interruptState?: AIInterruptState;
+}
